Capable of reading from wledb through express routes
