+++ backend/middleware/auth.js
@@ -1,3 +1,3 @@
 // Copyright (c) 2019 WingGroup
 
-//USED FOR USER SESSION AUTHETICATION
+//USED FOR USER SESSION AUTHENTICATION
